+++ spring-cloud-aws-samples/spring-cloud-aws-cloud-map-sample/src/main/resources/bootstrap.properties
@@ -6,8 +6,8 @@ aws.cloudmap.discovery.failFast=true
 aws.cloudmap.discovery.discoveryList[0].healthCheckProtocol=http
 aws.cloudmap.discovery.discoveryList[0].healthCheckResourcePath=/health
 aws.cloudmap.discovery.discoveryList[0].healthCheckThreshold=5
-aws.cloudmap.discovery.discoveryList[0].service=test-service
-aws.cloudmap.discovery.discoveryList[0].serviceNameSpace=test-namespace
+aws.cloudmap.discovery.discoveryList[0].service=hari-service
+aws.cloudmap.discovery.discoveryList[0].serviceNameSpace=hari-namespace
 
 # Register new instance
 aws.cloudmap.registry.description=Namespace for sample cloudmap registry service
