--- spring-cloud-starter-aws-cloudmap-config/pom.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>io.awspring.cloud</groupId>
-		<artifactId>spring-cloud-aws</artifactId>
-		<version>2.3.0-RC2</version>
-	</parent>
-
-	<artifactId>spring-cloud-starter-aws-cloudmap-config</artifactId>
-	<name>Spring Cloud AWS Cloud Map Configuration Starter</name>
-	<description>Spring Cloud AWS Cloud Map Configuration Starter</description>
-	<url>https://projects.spring.io/spring-cloud</url>
-	<organization>
-		<name>Pivotal Software, Inc.</name>
-		<url>https://www.spring.io</url>
-	</organization>
-	<properties>
-		<main.basedir>${basedir}/../..</main.basedir>
-	</properties>
-	<dependencies>
-		<!-- note that we don't depend on spring-cloud-aws-autoconfigure: that module brings in
-		     spring-cloud-aws-context and unwanted auto-configuration when you just want to use
-		     the Parameter Store configuration support.
-		     We simply include our own auto-configuration for Spring Cloud in this starter instead. -->
-		<dependency>
-			<groupId>io.awspring.cloud</groupId>
-			<artifactId>spring-cloud-aws-cloudmap</artifactId>
-			<version>2.3.0-RC2</version>
-		</dependency>
-		<dependency>
-			<groupId>io.awspring.cloud</groupId>
-			<artifactId>spring-cloud-aws-core</artifactId>
-		</dependency>
-	</dependencies>
-
-</project>
