+++ spring-cloud-starter-aws-cloudmap/src/main/resources/META-INF/spring.factories
